// src/app/generated/model/Program.ts
/**
 * JSBIS Services
 * API services for JSCBIS
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Program {
    id?: number;

    programId?: string;

    programName?: string;

    poc?: string;

    programManager?: string;

    descr?: string;

    notes?: string;

    functionalArea?: models.Label;

    medicalCategory?: models.Label;

    nbcCategory?: models.Label;

    coreCapability?: models.Label;

    primaryCapability?: models.Label;

    secondaryCapability?: models.Label;

    tertiaryCapability?: models.Label;

}